S08P12A409-277 Feat : 채점 완료 된 점수 반환 기능 추가

+++ Backend/src/main/java/com/ddockddack/domain/similarity/service/EnsembleModel.java
@@ -10,7 +10,7 @@ public class EnsembleModel {
 //    InputStream once ended, cannot reuse.
 //    Image data input format from frontend: ByteArray
 //    public static double CalculateSimilarity(InputStream inputStream1, InputStream inputStream2) throws Exception {
-    public static double CalculateSimilarity(byte[] byteArray1, byte[] byteArray2) throws Exception {
+    public static int CalculateSimilarity(byte[] byteArray1, byte[] byteArray2) throws Exception {
 
         final PerceptualHash perceptualHash = new PerceptualHash();
         final FeatureDetectorDescriptor featureDetectorDescriptor = new FeatureDetectorDescriptor();
